fix: Resolve Toast 500 error and task creation integer overflow
✅ FIXES RESOLVED:
1. Toast.tsx 500 error - Missing @radix-ui/react-toast dependency
   - Package installed and optimized by Vite
   - Added Chrome extension interference prevention

2. Task creation integer overflow - order field too small for timestamps
   - Changed Task.order from INTEGER to BIGINT in models
   - Applied database migration for existing data
   - Tested with actual French content that was failing

✅ TECHNICAL CHANGES:
- Added @radix-ui/react-toast to container dependencies
- Updated SQLAlchemy model: order = Column(BigInteger, default=0)
- Added Alembic migration: 61880637b235_change_task_order_field_to_biginteger
- Added meta tags in index.html to prevent extension interference
- Enhanced Toast component with accessibility attributes

✅ VALIDATION:
- Task creation API working: "Rechercher les forces de Grok4" ✓
- Toast notifications working without 500 errors ✓
- TypeScript compilation clean ✓
- Frontend loads successfully ✓

+++ app/models/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Text, ForeignKey, DateTime, Boolean, JSON
+from sqlalchemy import Column, Integer, BigInteger, String, Text, ForeignKey, DateTime, Boolean, JSON
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from app.db.config import Base
@@ -32,7 +32,7 @@ class Task(Base):
     title = Column(String, index=True, nullable=False)
     description = Column(Text, nullable=True) # Contenu de la tâche, éditable
     status = Column(String, default="À faire", nullable=False) # Ex: "À faire", "En cours", "Révision", "Terminé"
-    order = Column(Integer, default=0) # Pour l'assemblage
+    order = Column(BigInteger, default=0) # Pour l'assemblage
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), onupdate=func.now())
